regnet_app/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pandas==2.2.2
numpy==1.26.4
networkx==3.2.1
PyYAML==6.0.2
scipy==1.11.4
typing_extensions>=4.8.0
scikit-learn
matplotlib
seaborn
python-louvain
Flask
Jinja2